requirements.txt
google-generativeai
python-pptx
Pillow
requests